use float 16 for llama70b

+++ reproduction/tree_decoding.py
@@ -19,7 +19,7 @@ def get_gpu_usage():
     gpus = GPUtil.getGPUs()
     return gpus[0].memoryUsed
 
-dtype = torch.float
+dtype = torch.float16
 minFloat = torch.finfo(dtype).min
 device = "cuda" if torch.cuda.is_available() else "cpu"
 class SearchNode:
